+++ src/database/params/SiteParam.jl
@@ -333,10 +333,14 @@ function gc_to_comp_sites(param::AssocParam,sites::SiteParam)
                     j_gc,b_gc = site_translator_j[b]
                     #absolute index, relative to the Compressed4DMatrix
                     idx = validindex(ij_pair,a,b)
-                    if idx != 0 #if the index is valid
-                        ijab_val = param[i_gc,j_gc][a_gc,b_gc]
-                        if !_iszero(ijab_val) #if the value is not zero
-                            ij_pair.values.values[idx] =ijab_val
+                    if idx != 0
+                        ijab1 = param[i_gc,j_gc]
+                        idx_ijab = validindex(ijab1,a_gc,b_gc,false)
+                        if idx_ijab != 0
+                            ijab_val = ijab1.values.values[idx_ijab]
+                            if !_iszero(ijab_val) #if the value is not zero
+                                ij_pair.values.values[idx] =ijab_val
+                            end
                         end
                     end
                 end